Move all entity and enum definitions into single file
