--- Entities/Fields/FileField.cs
@@ -1,9 +1,0 @@
-namespace Keystone4Net.Entities;
-
-public class KeystoneFileField : KeystoneField
-{
-    public KeystoneFileField() : base("file")
-    {
-    }
-    public string Storage { get; set; } = string.Empty;
-}
