--- Entities/Fields/ImageField.cs
@@ -1,9 +1,0 @@
-namespace Keystone4Net.Entities;
-
-public class KeystoneImageField : KeystoneField
-{
-    public KeystoneImageField() : base("image")
-    {
-    }
-    public string Storage { get; set; } = string.Empty;
-}
