--- Entities/IKeystone.cs
@@ -1,6 +1,0 @@
-namespace Keystone4Net.Entities;
-
-public interface IKeystone
-{
-    void ConfigureKeystone(KeystoneConfig config);
-}
