--- Entities/KeystoneConfig.cs
@@ -1,38 +1,0 @@
-using Keystone4Net.Enums;
-using Microsoft.EntityFrameworkCore;
-
-namespace Keystone4Net.Entities;
-
-public class KeystoneConfig : KeystoneJsFunctionPropArgCall
-{
-    internal KeystoneConfig(DbContext dbContext, string? baseDir) : base(KeystoneImportObjects.Core, "config", null)
-    {
-        this.Db = new(dbContext, baseDir);
-    }
-
-    internal KeystoneDb Db { get; }
-
-    public KeystoneTypesConfig? Types { get; set; }
-
-    public KeystoneGraphqlOptions? Graphql { get; set; }
-
-    public KeystoneServerOptions? Server { get; set; }
-
-    public KeystoneSession? Session { get; set; }
-
-    public KeystoneUiSettings? Ui { get; set; }
-
-    public KeystoneJsFunction? ExtendGraphqlSchema { get; set; }
-
-    public Dictionary<string, KeystoneStorageConfig>? Storage { get; set; }
-
-    public bool Telemetry { get; set; }
-
-    public Dictionary<string, KeystoneList> Lists { get; } = [];
-
-    public void Add<T>(KeystoneList<T> value)
-    {
-        var typeName = typeof(T).Name;
-        this.Lists.Add(typeName, value);
-    }
-}
