--- Enums/KeystoneCookieSameSite.cs
@@ -1,10 +1,0 @@
-namespace Keystone4Net.Enums;
-
-public enum KeystoneCookieSameSite
-{
-    True,
-    False,
-    Lax,
-    Strict,
-    None
-}
