--- Enums/KeystoneDbProvider.cs
@@ -1,8 +1,0 @@
-namespace Keystone4Net.Enums;
-
-public enum KeystoneDbProvider
-{
-    Sqlite,
-    Postgresql,
-    Mysql
-}
