--- Enums/KeystoneDisplayMode.cs
@@ -1,10 +1,0 @@
-namespace Keystone4Net.Enums;
-
-public enum KeystoneDisplayMode
-{
-    Input,
-    Textarea,
-    Select,
-    Cards,
-    Count
-}
