--- Enums/KeystoneFieldMode.cs
@@ -1,9 +1,0 @@
-namespace Keystone4Net.Enums;
-
-public enum KeystoneFieldMode
-{
-    Edit,
-    Read,
-    Hidden
-}
-
