--- Enums/KeystoneImportObjects.cs
@@ -1,9 +1,0 @@
-namespace Keystone4Net.Enums;
-
-public enum KeystoneImportObjects
-{
-    Core,
-    Access,
-    Fields,
-    Session
-}
