--- Enums/KeystoneIndex.cs
@@ -1,8 +1,0 @@
-namespace Keystone4Net.Enums;
-
-public enum KeystoneIndex
-{
-    None,
-    Indexed,
-    Unique
-}
