--- Enums/KeystoneRemoveMode.cs
@@ -1,7 +1,0 @@
-namespace Keystone4Net.Enums;
-
-public enum KeystoneRemoveMode
-{
-    Disconnect,
-    None
-}
