--- Enums/KeystoneSortDirection.cs
@@ -1,8 +1,0 @@
-namespace Keystone4Net.Enums;
-
-public enum KeystoneSortDirection
-{
-    Asc,
-    Desc
-}
-
